pyproject.toml
[project]
name = "linkedin-layoff-scraper"
version = "0.1.0"
description = "A LinkedIn scraper for finding layoff posts"
authors = [
    {name = "Vishal Parekh",email = "vishalcparekh@outlook.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "langchain-openai>=0.0.5",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "browser-use>=0.1.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [
    { include = "src" }
]
